clean up commands code to match existing queue patterns
